First push
Stable, but each line is parsed as different tokens, which makes
replacing values not possible. For example,
"RN 01 02 5m" -> RN should be the name of the token, 01 and 02 should be
recognized as nodes, and the value should be 5m. Currently, RN, 01, 02,
and 5m are their own tokens, which is the easier approach, but not truly
a strictly correct parse